requirements.txt
streamlit
streamlit-drawable-canvas
numpy
pillow
scikit-learn
joblib
matplotlib
seaborn
